Make banner logic to util function instead of a hook
